images_service/requirements.txt
flask
flask-jwt-extended
werkzeug
pillow
flask_sqlalchemy
psycopg2-binary